pyproject.toml
[tool.poetry]
name = "schedule-software"
version = "0.1.0"
description = "Minimal Django + Create React App template"
authors = ["kevinshen56714 <kshen64@gatech.edu>"]
license = "MIT"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
